+++ src/pages/calendar/styles/calendar.module.scss
@@ -12,94 +12,6 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    // &__modal {
-    //   display: flex;
-    //   flex-direction: column;
-    //   position: absolute;
-    //   top: 10rem;
-    //   width: 80%;
-    //   height: 40%;
-    //   border-radius: 20px;
-    //   overflow: hidden;
-    //   z-index: 3;
-
-    //   background-color: white;
-    //   color: $color-purple-500;
-    //   text-align: center;
-
-    //   &__header {
-    //     display: flex;
-    //     flex-direction: row;
-    //     justify-content: space-between;
-    //     align-items: center;
-
-    //     padding: 1rem;
-    //     background-color: $color-purple-300;
-
-    //     &__date {
-    //       display: flex;
-    //       flex-direction: column;
-    //       justify-content: center;
-    //     }
-
-    //     &__icon {
-    //       cursor: pointer;
-    //     }
-    //   }
-    //   &__body {
-    //     // overflow: scroll;
-    //     height: 100%;
-    //     &__event {
-    //       width: 100%;
-    //       height: 20%;
-    //       border-bottom: 1px solid $color-gray-200;
-    //       padding: 1rem;
-
-    //       display: flex;
-    //       flex-direction: row;
-    //       justify-content: space-between;
-    //       align-items: center;
-
-    //       &__icon {
-    //         cursor: pointer;
-    //       }
-
-    //       &.election {
-    //         color: red;
-    //         cursor: pointer;
-    //       }
-    //       &.custom {
-    //         color: $color-purple-500;
-    //       }
-    //     }
-    //     &__inputBox {
-    //       width: 100%;
-    //       height: 3rem;
-    //       gap: 5px;
-    //       padding-left: 5px;
-    //       padding-right: 5px;
-
-    //       display: flex;
-    //       flex-direction: row;
-    //       justify-content: center;
-    //       align-items: center;
-    //       &__inputTag {
-    //         width: 80%;
-    //         height: 2rem;
-    //         border-radius: 10px;
-    //       }
-    //       &__writeButton {
-    //         width: 20%;
-    //         height: 2rem;
-    //         border-radius: 10px;
-    //         cursor: pointer;
-
-    //         background-color: $color-purple-500;
-    //         color: white;
-    //       }
-    //     }
-    //   }
-    // }
 
     &__topBar {
       width: 100%;
