+++ src/pages/calendar/topbar.module.scss
@@ -1,0 +1,14 @@
+.topBar {
+  width: 100%;
+  height: 4rem;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 5px;
+
+  background-color: $color-purple-300;
+
+  &__prevArrow {
+    cursor: pointer;
+  }
+}
